+++ ocgapi.cpp
@@ -86,7 +86,7 @@ extern "C" DECL_DLLEXPORT void start_duel(ptr pduel, int options) {
 			pd->game_field->player[0].tag_list_hand.push_back(pcard);
 			pcard->current.controler = 0;
 			pcard->current.location = LOCATION_HAND;
-			pcard->current.sequence = pd->game_field->player[0].tag_list_hand.size() - 1;
+			pcard->current.sequence = (uint8)pd->game_field->player[0].tag_list_hand.size() - 1;
 			pcard->current.position = POS_FACEDOWN;
 		}
 		for(int i = 0; i < pd->game_field->player[1].start_count && pd->game_field->player[1].tag_list_main.size(); ++i) {
@@ -95,7 +95,7 @@ extern "C" DECL_DLLEXPORT void start_duel(ptr pduel, int options) {
 			pd->game_field->player[1].tag_list_hand.push_back(pcard);
 			pcard->current.controler = 1;
 			pcard->current.location = LOCATION_HAND;
-			pcard->current.sequence = pd->game_field->player[1].tag_list_hand.size() - 1;
+			pcard->current.sequence = (uint8)pd->game_field->player[1].tag_list_hand.size() - 1;
 			pcard->current.position = POS_FACEDOWN;
 		}
 	}
@@ -160,7 +160,7 @@ extern "C" DECL_DLLEXPORT void new_tag_card(ptr pduel, uint32 code, uint8 owner,
 		pcard->owner = owner;
 		pcard->current.controler = owner;
 		pcard->current.location = LOCATION_DECK;
-		pcard->current.sequence = ptduel->game_field->player[owner].tag_list_main.size() - 1;
+		pcard->current.sequence = (uint8)ptduel->game_field->player[owner].tag_list_main.size() - 1;
 		pcard->current.position = POS_FACEDOWN_DEFENSE;
 		break;
 	case LOCATION_EXTRA:
@@ -168,7 +168,7 @@ extern "C" DECL_DLLEXPORT void new_tag_card(ptr pduel, uint32 code, uint8 owner,
 		pcard->owner = owner;
 		pcard->current.controler = owner;
 		pcard->current.location = LOCATION_EXTRA;
-		pcard->current.sequence = ptduel->game_field->player[owner].tag_list_extra.size() - 1;
+		pcard->current.sequence = (uint8)ptduel->game_field->player[owner].tag_list_extra.size() - 1;
 		pcard->current.position = POS_FACEDOWN_DEFENSE;
 		break;
 	}
@@ -298,7 +298,7 @@ extern "C" DECL_DLLEXPORT int32 query_field_info(ptr pduel, byte* buf) {
 			if(pcard) {
 				*buf++ = 1;
 				*buf++ = pcard->current.position;
-				*buf++ = pcard->xyz_materials.size();
+				*buf++ = (uint8)pcard->xyz_materials.size();
 			} else {
 				*buf++ = 0;
 			}
@@ -312,14 +312,14 @@ extern "C" DECL_DLLEXPORT int32 query_field_info(ptr pduel, byte* buf) {
 				*buf++ = 0;
 			}
 		}
-		*buf++ = ptduel->game_field->player[playerid].list_main.size();
-		*buf++ = ptduel->game_field->player[playerid].list_hand.size();
-		*buf++ = ptduel->game_field->player[playerid].list_grave.size();
-		*buf++ = ptduel->game_field->player[playerid].list_remove.size();
-		*buf++ = ptduel->game_field->player[playerid].list_extra.size();
+		*buf++ = (uint8)ptduel->game_field->player[playerid].list_main.size();
+		*buf++ = (uint8)ptduel->game_field->player[playerid].list_hand.size();
+		*buf++ = (uint8)ptduel->game_field->player[playerid].list_grave.size();
+		*buf++ = (uint8)ptduel->game_field->player[playerid].list_remove.size();
+		*buf++ = (uint8)ptduel->game_field->player[playerid].list_extra.size();
 		*buf++ = ptduel->game_field->player[playerid].extra_p_count;
 	}
-	*buf++ = ptduel->game_field->core.current_chain.size();
+	*buf++ = (uint8)ptduel->game_field->core.current_chain.size();
 	for(auto chit = ptduel->game_field->core.current_chain.begin(); chit != ptduel->game_field->core.current_chain.end(); ++chit) {
 		effect* peffect = chit->triggering_effect;
 		*((int*)(buf)) = peffect->handler->data.code;
